+++ addons/gut/test.gd
@@ -145,6 +145,8 @@ var _strutils = _utils.Strutils.new()
 # syntax sugar
 var ParameterFactory = _utils.ParameterFactory
 var CompareResult = _utils.CompareResult
+var InputFactory = _utils.InputFactory
+var InputSender = _utils.InputSender
 
 func _init():
 	DOUBLE_STRATEGY = _utils.DOUBLE_STRATEGY # yes, this is right
@@ -347,6 +349,7 @@ func assert_ne(got, not_expected, text=""):
 			_fail(disp)
 		else:
 			_pass(disp)
+
 
 # ------------------------------------------------------------------------------
 # Asserts that the expected value almost equals the value got.
@@ -1006,7 +1009,7 @@ func assert_not_null(got, text=''):
 # Asserts object has been freed from memory
 # We pass in a title (since if it is freed, we lost all identity data)
 # -----------------------------------------------------------------------------
-func assert_freed(obj, title):
+func assert_freed(obj, title='something'):
 	var disp = title
 	if(is_instance_valid(obj)):
 		disp = _strutils.type2str(obj) + title
@@ -1077,6 +1080,7 @@ func _validate_assert_setget_called_input(type, name_property
 
 	obj.free()
 	return result
+
 
 # ------------------------------------------------------------------------------
 # Asserts the given setter and getter methods are called when the given property
@@ -1200,6 +1204,10 @@ func pending(text=""):
 		_lgr.pending(text)
 		gut._pending(text)
 
+# ------------------------------------------------------------------------------
+# Returns the number of times a signal was emitted.  -1 returned if the object
+# is not being watched.
+# ------------------------------------------------------------------------------
 
 # ------------------------------------------------------------------------------
 # Yield for the time sent in.  The optional message will be printed when
@@ -1363,6 +1371,7 @@ func double_inner(path, subpath, strategy=null):
 	var override_strat = _utils.nvl(strategy, gut.get_doubler().get_strategy())
 	return gut.get_doubler().double_inner(path, subpath, override_strat)
 
+
 # ------------------------------------------------------------------------------
 # Add a method that the doubler will ignore.  You can pass this the path to a
 # script or scene or a loaded script or scene.  When running tests, these
@@ -1378,7 +1387,6 @@ func ignore_method_when_doubling(thing, method_name):
 			path = inst.get_script().get_path()
 
 	gut.get_doubler().add_ignored_method(path, method_name)
-
 
 # ------------------------------------------------------------------------------
 # Stub something.